requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-heroku
gunicorn
django-cors-headers
django-filter==2.4.0 
python-dateutil==2.8.1
pandas==1.2.4